Use server to return tokens rather than hard coding secrets
Secrets were hard coded, specifically for Twilio. Since we don't want
these in source control and can't live in the app itself, we need a
backend server to store the secrets in and use them to retrieve the
token needed to instantiate the chat client. This adds the server but it
is not yet used in the app.