Use local scope for bash variables
Variable definitions in functions have global scope. We want them to
only be scoped locally to the function.

Also in run.sh the vars are given an underscore prefix because they are
globally scoped and can't be locally scoped since they are not in a
function.

Also to note that if you scope variables locally to a function, they are
still scoped to functions in which it calls.