Don't clear typed message when a message is received
Users who typed in a message would have their text are cleared if a
message was received. The intent was to have their message cleared when
they sent a message because them sending a message causes the channel to
trigger the "messageAdded" event which clears the text area. However,
this also happens when the other user sends a message so it's not
desirable. This fix causes the text are to be cleared when the user
sends a message rather than one is received.