Refactor channel joining functionality
Pull channel joining functionality into it's own file.

The main reason for this change is because calling joinChannel() from
useEffect was throwing a warning saying that it needed to be added as a
dependency. However, doing so caused it to run in a loop, constantly
rendering. I'm not exactly sure why this was but it didn't seem right to
do anyways. Instead, the function is now being pulled from another
class. I'm not sure why it is now no longer required to be passed in as
a dependency though. Perhaps because now we can no longer access global
variables like we were. But passing these values in rather than global
access is better design anyways.

Side note that this started causing failures in the build because we
recently started using `npm ci` rather than `npm install` for the build.
The former fails if there are warnings.